Add files via upload
HLS4 version was never finished by the end of SWENG861.  I spoke with Professor Rodini, and he commented that in this version there was a syntax error in the working program that circumvented polymorphism.  After the deadline he later commented that he may have thrown out a "red herring" on accident.  This version was never submitted as time ran out and HLS3.1 met all of the requirements for the assignment (sans stretch goals).